[011] - Verificação do usuario logado nas açoes do sistema.

+++ src/main/java/br/com/github/macgarcia/appanotacao/controllers/AnotacaoController.java
@@ -20,6 +20,8 @@ import br.com.github.macgarcia.appanotacao.repositorys.AnotacaoRepository;
 @Controller
 public class AnotacaoController {
 	
+	private static final String CAMINHO_INICIO_APP = "login/index";
+	
 	private static final String CAMINHO_PAGINA_ANOTACAO = "anotacao/anotacao-index";
 	private static final String CAMINHO_PAGINA_NOVO_EDITAR = "anotacao/novo-editar-anotacao.html";
 	
@@ -38,7 +40,11 @@ public class AnotacaoController {
 		
 		setIndexController(indexController);
 		
-		usuarioLogado = (Usuario) session.getAttribute("Usuario");
+		final boolean podeUsarSistema = verificarUsoDoSistema(session);
+		
+		if (!podeUsarSistema) {
+			return new ModelAndView(CAMINHO_INICIO_APP);
+		}
 		
 		if (!executouPesquisa) {
 			alimentarListaAnotacoes("");
@@ -53,6 +59,12 @@ public class AnotacaoController {
 	
 	@PostMapping(path = "/pesquisa")
 	public ModelAndView executarPesquisa(final HttpSession session, final String pesquisa) {
+		
+		final boolean podeUsarSistema = verificarUsoDoSistema(session);
+		if (!podeUsarSistema) {
+			return new ModelAndView(CAMINHO_INICIO_APP);
+		}
+		
 		this.executouPesquisa = true;
 		alimentarListaAnotacoes(pesquisa);
 		return telaDeAnotacoes(session, null);
@@ -70,6 +82,13 @@ public class AnotacaoController {
 	
 	@GetMapping(path = "/novo-editar")
 	public ModelAndView novoEditar(final HttpSession session, final Long id) {
+		
+		final boolean podeUsarSistema = verificarUsoDoSistema(session);
+		
+		if (!podeUsarSistema) {
+			return new ModelAndView(CAMINHO_INICIO_APP);
+		}
+		
 		final ModelAndView model = new ModelAndView(CAMINHO_PAGINA_NOVO_EDITAR);
 		String titulo = null;
 		if (id != 0) {
@@ -89,6 +108,12 @@ public class AnotacaoController {
 	@PostMapping(path = "/salvar")
 	@Transactional
 	public ModelAndView atualizar(final HttpSession session, final Anotacao anotacao) {
+		
+		final boolean podeUsarSistema = verificarUsoDoSistema(session);
+		
+		if (!podeUsarSistema) {
+			return new ModelAndView(CAMINHO_INICIO_APP);
+		}
 		anotacao.setUsuario(usuarioLogado);
 		repository.saveAndFlush(anotacao);
 		return telaDeAnotacoes(session, null);
@@ -97,6 +122,13 @@ public class AnotacaoController {
 	@GetMapping(path = "/apagar")
 	@Transactional
 	public ModelAndView apagar(final HttpSession session, final Long id) {
+		
+		final boolean podeUsarSistema = verificarUsoDoSistema(session);
+		
+		if (!podeUsarSistema) {
+			return new ModelAndView(CAMINHO_INICIO_APP);
+		}
+		
 		repository.deleteById(id);
 		return telaDeAnotacoes(session, null);
 	}
@@ -130,6 +162,14 @@ public class AnotacaoController {
 			count++;
 		}
 		return aEncontrada;
+	}
+	
+	private boolean verificarUsoDoSistema(final HttpSession session) {
+		final Usuario usuario = (Usuario) session.getAttribute("Usuario");
+		if (usuario == null) {
+			return false;
+		}
+		return true;
 	}
 	
 }
